requirements.txt
Flask==0.12.2
Flask-Cors==3.0.3
PyJWT==1.5.2
marshmallow==2.13.6
marshmallow-mongoengine==0.7.8
mongoengine==0.13.0
